the new master data
you can use CTVControl to perform basic auth and show mouse.
CTV Control handle static methods required in order to send real remote
command to tv ( http and udp )
the http post request is implemented now ( for auth ( completed ) and
show mouse on tv(on eventmanager ), but need to re-code EventManager ,
commandManager and Query functions.
It is not required at this time the feedback listener (http and udp )
UDPRequest need to implement real UDP send-packet-to-tv.

An issue that need to be fixed: when cursor is shown on tv, a qmenu also
appear. need to reread java sources to see how to fix ;)

A console test program will be posted soon ;)